switched integer to decimal

+++ OrderProcessingSystem.Models/Orders.cs
@@ -12,7 +12,7 @@ namespace OrderProcessingSystem.Models
 
         [Required]
         public string OrderDetails { get; set; }
-        public int AmountDue { get; set; }
+        public decimal AmountDue { get; set; }
         public bool IsComplete { get; set; }
     }
 }
